// src/app/progress/models/progress-statistics.ts
export class ProgressStatistics {
  "pointsScoredToYear": number[];
  "exercisesPerformedToYear": number[];
  "exercisesCompletedToYear": number[];
  "pointsScoredToMonth": number[];
  "exercisesPerformedToMonth": number[];
  "exercisesCompletedToMonth": number[];
  "pointsScoredToHour": number[];
  "exercisesPerformedToHour": number[];
  "exercisesCompletedToHour": number[];
}